Day03 - Part 1  & 2

+++ src/main/kotlin/xyz/coincoin/adventofcode2020/Day.kt
@@ -1,6 +1,6 @@
 package xyz.coincoin.adventofcode2020
 
 interface Day {
-    fun part1(input: List<String>): Int
-    fun part2(input: List<String>): Int
+    fun part1(input: List<String>): Long
+    fun part2(input: List<String>): Long
 }
